refactor: separate projects from info in routes, controllers and tests
