+++ src/api/index.ts
@@ -1,6 +1,8 @@
 import { Router, type Request, type Response, type NextFunction } from "express";
 
 import infoRoutes from "./routes/info.js";
+import projectRoutes from "./routes/project.js";
+
 import { AppError } from "./models/appError.js";
 
 import swaggerUi from "swagger-ui-express";
@@ -9,6 +11,7 @@ import swagger from "../docs/swagger.json";
 const apiRouter = Router();
 
 apiRouter.use("/", infoRoutes);
+apiRouter.use("/projects", projectRoutes);
 
 apiRouter.use("/docs", swaggerUi.serve, swaggerUi.setup(swagger));
